implement linux hooks
