+++ lib/file_system.py
@@ -4,8 +4,8 @@ if sys.platform == 'darwin':
     from ..lib.platform.darwin.file_system import *
 elif sys.platform == 'win32':
     from ..lib.platform.win32.file_system import *
-elif sys.platform == 'linux2':
-    from ..lib.platform.linux2.file_system import *
+elif sys.platform in ('linux', 'linux2'):
+    from ..lib.platform.linux.file_system import *
 else:
     from ..lib.platform.unsupported.file_system import *
