+++ lib/link_opener.py
@@ -4,8 +4,8 @@ if sys.platform == 'darwin':
     from ..lib.platform.darwin.link_opener import *
 elif sys.platform == 'win32':
     from ..lib.platform.win32.link_opener import *
-elif sys.platform == 'linux2':
-    from ..lib.platform.linux2.link_opener import *
+elif sys.platform in ('linux', 'linux2'):
+    from ..lib.platform.linux.link_opener import *
 else:
     from ..lib.platform.unsupported.link_opener import *
