+++ lib/platform/linux/file_system.py
@@ -1,4 +1,4 @@
 __all__ = ['_path_for_url']
 
 def _path_for_url(path):
-    return None
+    return path.replace('/', ':')
