--- lib/platform/linux2/app_controller.py
@@ -1,13 +1,0 @@
-__all__ = ['_launch_kite', '_locate_kite', '_is_kite_running']
-
-
-def _launch_kite(app):
-    pass
-
-
-def _locate_kite():
-    return False, None
-
-
-def _is_kite_running():
-    return False
