--- lib/platform/linux2/link_opener.py
@@ -1,18 +1,0 @@
-import subprocess
-
-__all__ = [
-    '_open_browser', '_open_browser_url',
-    '_open_copilot', '_open_copilot_root'
-]
-
-def _open_browser(ident):
-    return None
-
-def _open_browser_url(url):
-    return None
-
-def _open_copilot(ident):
-    return None
-
-def _open_copilot_root(path):
-    return None
